split logger code into multiple files
